# Conflicts:
#	plugin-azure-server/src/main/kotlin/jetbrains/buildServer/clouds/azure/arm/AzureCloudClientFactory.kt
#	plugin-azure-server/src/main/kotlin/jetbrains/buildServer/clouds/azure/arm/AzureConstants.kt

+++ plugin-azure-server/src/main/kotlin/jetbrains/buildServer/clouds/azure/arm/AzureCloudImage.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2020 JetBrains s.r.o.
+ * Copyright 2000-2021 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,33 +17,46 @@
 package jetbrains.buildServer.clouds.azure.arm
 
 import com.intellij.openapi.diagnostic.Logger
+import jetbrains.buildServer.clouds.CanStartNewInstanceResult
 import jetbrains.buildServer.clouds.CloudException
 import jetbrains.buildServer.clouds.CloudInstanceUserData
 import jetbrains.buildServer.clouds.InstanceStatus
 import jetbrains.buildServer.clouds.QuotaException
 import jetbrains.buildServer.clouds.azure.AzureUtils
 import jetbrains.buildServer.clouds.azure.arm.connector.AzureApiConnector
-import jetbrains.buildServer.clouds.azure.arm.types.*
+import jetbrains.buildServer.clouds.azure.arm.connector.AzureInstance
+import jetbrains.buildServer.clouds.azure.arm.types.AzureContainerHandler
+import jetbrains.buildServer.clouds.azure.arm.types.AzureHandler
+import jetbrains.buildServer.clouds.azure.arm.types.AzureImageHandler
+import jetbrains.buildServer.clouds.azure.arm.types.AzureInstanceHandler
+import jetbrains.buildServer.clouds.azure.arm.types.AzureTemplateHandler
+import jetbrains.buildServer.clouds.azure.arm.types.AzureVhdHandler
 import jetbrains.buildServer.clouds.base.AbstractCloudImage
 import jetbrains.buildServer.clouds.base.connector.AbstractInstance
 import jetbrains.buildServer.clouds.base.errors.TypedCloudErrorInfo
 import jetbrains.buildServer.serverSide.TeamCityProperties
-import kotlinx.coroutines.*
-import java.lang.StringBuilder
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import java.util.*
+import java.util.concurrent.atomic.AtomicReference
 
 /**
  * Azure cloud image.
  */
-class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDetails,
-                                  private val myApiConnector: AzureApiConnector)
-    : AbstractCloudImage<AzureCloudInstance, AzureCloudImageDetails>(myImageDetails.sourceId, myImageDetails.sourceId) {
+class AzureCloudImage(
+    private val myImageDetails: AzureCloudImageDetails,
+    private val myApiConnector: AzureApiConnector,
+    private val myScope: CoroutineScope
+) : AbstractCloudImage<AzureCloudInstance, AzureCloudImageDetails>(myImageDetails.sourceId, myImageDetails.sourceId) {
 
     private val myImageHandlers = mapOf(
-            AzureCloudImageType.Vhd to AzureVhdHandler(myApiConnector),
-            AzureCloudImageType.Image to AzureImageHandler(myApiConnector),
-            AzureCloudImageType.Template to AzureTemplateHandler(myApiConnector),
-            AzureCloudImageType.Container to AzureContainerHandler(myApiConnector)
+        AzureCloudImageType.Vhd to AzureVhdHandler(myApiConnector),
+        AzureCloudImageType.Image to AzureImageHandler(myApiConnector),
+        AzureCloudImageType.GalleryImage to AzureImageHandler(myApiConnector),
+        AzureCloudImageType.Template to AzureTemplateHandler(myApiConnector),
+        AzureCloudImageType.Container to AzureContainerHandler(myApiConnector)
     )
     private val myInstanceHandler = AzureInstanceHandler(myApiConnector)
 
@@ -58,6 +71,8 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
     )
 
     private var azureCpuQuotaExceeded: Set<String>? = null
+
+    private val overlimitInstanceToDelete = AtomicReference<AzureCloudInstance?>(null)
 
     override fun getImageDetails(): AzureCloudImageDetails = myImageDetails
 
@@ -81,25 +96,27 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
         }
     }
 
-    override fun canStartNewInstance(): Boolean {
-        if (activeInstances.size >= myImageDetails.maxInstances) return false
+    override fun canStartNewInstance(): CanStartNewInstanceResult {
+        if (activeInstances.size >= myImageDetails.maxInstances) {
+            return CanStartNewInstanceResult.no("Quota exceeded: ActiveInstances count reached MaxInstances value of ${myImageDetails.maxInstances}")
+        }
         // Check Azure CPU quota state
         azureCpuQuotaExceeded?.let { instances ->
             if (instances == getInstanceIds()) {
-                return false
+                return CanStartNewInstanceResult.no("Azure CPU quota exceeded")
             } else {
                 azureCpuQuotaExceeded = null
                 LOG.info("Azure CPU quota limit has been reset due to change in the number of active instances for image ${imageDetails.sourceId}.")
             }
         }
         if (imageDetails.deployTarget == AzureCloudDeployTarget.Instance && stoppedInstances.isEmpty()) {
-            return false
+            return CanStartNewInstanceResult.no("Stopped Instances pool is empty, but we trying to run yet another instance. Check inactive agents for errors")
         }
-        return true
+        return CanStartNewInstanceResult.yes()
     }
 
     override fun startNewInstance(userData: CloudInstanceUserData) = runBlocking {
-        if (!canStartNewInstance()) {
+        if (!canStartNewInstance().isPositive) {
             throw QuotaException("Unable to start more instances. Limit has reached")
         }
 
@@ -120,23 +137,27 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
      * @return created instance.
      */
     private fun createInstance(userData: CloudInstanceUserData): AzureCloudInstance {
-        val name = getInstanceName()
-        val instance = AzureCloudInstance(this, name)
-        instance.status = InstanceStatus.SCHEDULED_TO_START
-        val data = AzureUtils.setVmNameForTag(userData, name)
+        val instance = createAzureCloudInstance()
+        val data = AzureUtils.setVmNameForTag(userData, instance.name)
+        val image = this
 
-        GlobalScope.launch {
+        myScope.launch {
             val hash = handler!!.getImageHash(imageDetails)
 
             instance.properties[AzureConstants.TAG_PROFILE] = userData.profileId
             instance.properties[AzureConstants.TAG_SOURCE] = imageDetails.sourceId
             instance.properties[AzureConstants.TAG_DATA_HASH] = getDataHash(data)
             instance.properties[AzureConstants.TAG_IMAGE_HASH] = hash
+            parseCustomTags(myImageDetails.customTags).forEach {
+                instance.properties[it.first] = it.second
+            }
 
             try {
+                instance.provisioningInProgress = true
+                instance.status = InstanceStatus.STARTING
                 LOG.info("Creating new virtual machine ${instance.name}")
                 myApiConnector.createInstance(instance, data)
-                instance.status = InstanceStatus.RUNNING
+                updateInstanceStatus(image, instance)
             } catch (e: Throwable) {
                 LOG.warnAndDebugDetails(e.message, e)
                 handleDeploymentError(e)
@@ -149,7 +170,6 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
                     try {
                         myApiConnector.deleteInstance(instance)
                         LOG.info("Allocated resources for virtual machine ${instance.name} have been removed")
-                        removeInstance(instance.instanceId)
                     } catch (e: Throwable) {
                         val message = "Failed to delete allocated resources for virtual machine ${instance.name}: ${e.message}"
                         LOG.warnAndDebugDetails(message, e)
@@ -157,12 +177,32 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
                 } else {
                     LOG.info("Allocated resources for virtual machine ${instance.name} would not be deleted. Cleanup them manually.")
                 }
+            } finally {
+                instance.provisioningInProgress = false
             }
         }
-
-        addInstance(instance)
-
         return instance
+    }
+
+    private fun createAzureCloudInstance(): AzureCloudInstance {
+        do {
+            val instance = AzureCloudInstance(this, getInstanceName())
+            instance.status = InstanceStatus.SCHEDULED_TO_START
+
+            if (addInstanceIfAbsent(instance)) {
+                while(activeInstances.size > myImageDetails.maxInstances) {
+                    instances.filter { it.status == InstanceStatus.SCHEDULED_TO_START }.maxByOrNull { it.name }
+                        ?: throw QuotaException("Unable to start more instances. Limit has reached")
+
+                    if (overlimitInstanceToDelete.compareAndSet(null, instance)) {
+                        removeInstance(instance.instanceId)
+                        overlimitInstanceToDelete.set(null)
+                        throw QuotaException("Unable to start more instances. Limit has reached")
+                    }
+                }
+                return instance
+            }
+        } while (true)
     }
 
     /**
@@ -170,13 +210,14 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
      *
      * @return instance if it found.
      */
-    private suspend fun tryToStartStoppedInstance(userData: CloudInstanceUserData) = coroutineScope {
-        val instances = stoppedInstances
-        if (instances.isNotEmpty()) {
-            val validInstances = if (myImageDetails.behaviour.isDeleteAfterStop) {
-                LOG.info("Will remove all virtual machines due to cloud image settings")
-                emptyList()
-            } else {
+    private suspend fun tryToStartStoppedInstance(userData: CloudInstanceUserData): AzureCloudInstance? {
+        val image = this
+        return coroutineScope {
+            if (myImageDetails.behaviour.isDeleteAfterStop && myImageDetails.spotVm != true) return@coroutineScope null
+            if (stoppedInstances.isEmpty()) return@coroutineScope null
+
+            val instances = stoppedInstances
+            val validInstances =
                 instances.filter {
                     if (!isSameImageInstance(it)) {
                         LOG.info("Will remove virtual machine ${it.name} due to changes in image source")
@@ -189,45 +230,51 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
                     }
                     return@filter true
                 }
-            }
 
-            val invalidInstances = instances - validInstances
+            val invalidInstances = instances - validInstances.toSet()
             val instance = validInstances.firstOrNull()
 
             instance?.status = InstanceStatus.SCHEDULED_TO_START
 
-            GlobalScope.launch {
+            myScope.launch {
                 invalidInstances.forEach {
-                    try {
-                        LOG.info("Removing virtual machine ${it.name}")
-                        myApiConnector.deleteInstance(it)
-                        removeInstance(it.instanceId)
-                    } catch (e: Throwable) {
-                        LOG.warnAndDebugDetails(e.message, e)
-                        it.status = InstanceStatus.ERROR
-                        it.updateErrors(TypedCloudErrorInfo.fromException(e))
+                    val instanceToRemove = removeInstance(it.instanceId)
+                    if (instanceToRemove != null) {
+                        try {
+                            LOG.info("Removing virtual machine ${it.name}")
+                            it.provisioningInProgress = true
+                            myApiConnector.deleteInstance(it)
+                        } catch (e: Throwable) {
+                            LOG.warnAndDebugDetails(e.message, e)
+                            it.status = InstanceStatus.ERROR
+                            it.updateErrors(TypedCloudErrorInfo.fromException(e))
+                            addInstance(instanceToRemove)
+                        } finally {
+                            it.provisioningInProgress = false
+                        }
                     }
                 }
 
                 instance?.let {
                     try {
+                        instance.provisioningInProgress = true
+                        it.status = InstanceStatus.STARTING
                         LOG.info("Starting stopped virtual machine ${it.name}")
                         myApiConnector.startInstance(it)
-                        instance.status = InstanceStatus.RUNNING
+                        updateInstanceStatus(image, it)
                     } catch (e: Throwable) {
                         LOG.warnAndDebugDetails(e.message, e)
                         handleDeploymentError(e)
 
                         it.status = InstanceStatus.ERROR
                         it.updateErrors(TypedCloudErrorInfo.fromException(e))
+                    } finally {
+                        instance.provisioningInProgress = false
                     }
                 }
             }
-
             return@coroutineScope instance
         }
-
-        null
     }
 
     /**
@@ -241,21 +288,36 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
 
         instance.status = InstanceStatus.SCHEDULED_TO_START
 
-        GlobalScope.launch {
+        val image = this
+        myScope.launch {
             try {
+                instance.provisioningInProgress = true
+                instance.status = InstanceStatus.STARTING
                 LOG.info("Starting virtual machine ${instance.name}")
                 myApiConnector.startInstance(instance)
-                instance.status = InstanceStatus.RUNNING
+                updateInstanceStatus(image, instance)
             } catch (e: Throwable) {
                 LOG.warnAndDebugDetails(e.message, e)
                 handleDeploymentError(e)
 
                 instance.status = InstanceStatus.ERROR
                 instance.updateErrors(TypedCloudErrorInfo.fromException(e))
+            } finally {
+                instance.provisioningInProgress = false
             }
         }
 
         return instance
+    }
+
+    private fun updateInstanceStatus(image: AzureCloudImage, instance: AzureCloudInstance) {
+        val instances = myApiConnector.fetchInstances<AzureInstance>(image)
+
+        instances[instance.name]?.let { azureInstance ->
+            azureInstance.startDate?.let { instance.setStartDate(it) }
+            azureInstance.ipAddress?.let { instance.setNetworkIdentify(it) }
+            instance.status = azureInstance.instanceStatus
+        }
     }
 
     private suspend fun isSameImageInstance(instance: AzureCloudInstance) = coroutineScope {
@@ -281,15 +343,20 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
     override fun restartInstance(instance: AzureCloudInstance) {
         instance.status = InstanceStatus.RESTARTING
 
-        GlobalScope.launch {
+        val image = this
+        myScope.launch {
             try {
+                instance.provisioningInProgress = true
+                instance.status = InstanceStatus.STARTING
                 LOG.info("Restarting virtual machine ${instance.name}")
                 myApiConnector.restartInstance(instance)
-                instance.status = InstanceStatus.RUNNING
+                updateInstanceStatus(image, instance)
             } catch (e: Throwable) {
                 LOG.warnAndDebugDetails(e.message, e)
                 instance.status = InstanceStatus.ERROR
                 instance.updateErrors(TypedCloudErrorInfo.fromException(e))
+            } finally {
+                instance.provisioningInProgress = false
             }
         }
     }
@@ -300,23 +367,26 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
             return
         }
 
+        val image = this
         instance.status = InstanceStatus.SCHEDULED_TO_STOP
 
-        GlobalScope.launch {
+        myScope.launch {
             try {
+                instance.provisioningInProgress = true
+                instance.status = InstanceStatus.STOPPING
                 val sameVhdImage = isSameImageInstance(instance)
                 if (myImageDetails.behaviour.isDeleteAfterStop) {
                     LOG.info("Removing virtual machine ${instance.name} due to cloud image settings")
                     myApiConnector.deleteInstance(instance)
-                    removeInstance(instance.instanceId)
+                    instance.status = InstanceStatus.STOPPED
                 } else if (!sameVhdImage) {
                     LOG.info("Removing virtual machine ${instance.name} due to cloud image retention policy")
                     myApiConnector.deleteInstance(instance)
-                    removeInstance(instance.instanceId)
+                    instance.status = InstanceStatus.STOPPED
                 } else {
                     LOG.info("Stopping virtual machine ${instance.name}")
                     myApiConnector.stopInstance(instance)
-                    instance.status = InstanceStatus.STOPPED
+                    updateInstanceStatus(image, instance)
                 }
 
                 LOG.info("Virtual machine ${instance.name} has been successfully terminated")
@@ -324,6 +394,8 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
                 LOG.warnAndDebugDetails(e.message, e)
                 instance.status = InstanceStatus.ERROR
                 instance.updateErrors(TypedCloudErrorInfo.fromException(e))
+            } finally {
+                instance.provisioningInProgress = false
             }
         }
     }
@@ -338,8 +410,8 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
         }
 
     private fun getInstanceName(): String {
-        val keys = instances.map { it.instanceId.toLowerCase() }
-        val sourceName = myImageDetails.sourceId.toLowerCase()
+        val keys = instances.map { it.instanceId.lowercase(Locale.getDefault()) }
+        val sourceName = myImageDetails.sourceId.lowercase(Locale.getDefault())
         var i = 1
 
         while (keys.contains(sourceName + i)) i++
@@ -378,5 +450,21 @@ class AzureCloudImage constructor(private val myImageDetails: AzureCloudImageDet
     companion object {
         private val LOG = Logger.getInstance(AzureCloudImage::class.java.name)
         private val AZURE_CPU_QUOTA_EXCEEDED = Regex("Operation results in exceeding quota limits of Core\\. Maximum allowed: \\d+, Current in use: \\d+, Additional requested: \\d+\\.")
+
+        fun parseCustomTags(rawString: String?): List<Pair<String, String>> {
+            return if (!rawString.isNullOrEmpty()) {
+                rawString.lines().map { it.trim() }.filter { it.isNotEmpty() }.mapNotNull {
+                    val tag = it
+                    val equalsSignIndex = tag.indexOf("=")
+                    if (equalsSignIndex >= 1) {
+                        Pair(tag.substring(0, equalsSignIndex), tag.substring(equalsSignIndex + 1))
+                    } else {
+                        null
+                    }
+                }
+            } else {
+                Collections.emptyList()
+            }
+        }
     }
 }
